+++ Modules/ShortUrl/resources/views/user/edit.blade.php
@@ -398,4 +398,9 @@
     </div>
 </form>
 
+
+{{-- Nettoyage des blancs invisibles colles autour des URL (demande 2026-09-05).
+     Le serveur normalise deja (trait NormalizesPastedUrls) : ceci rend le
+     nettoyage VISIBLE dans le champ. --}}
+@include('shorturl::partials._url-paste-clean')
 @endsection